+++ lib/src/elements/source/shared_texture/windows/mod.rs
@@ -1,0 +1,5 @@
+mod d3d11_shared_texture_source;
+
+pub use d3d11_shared_texture_source::{
+    D3d11SharedTextureHandle, D3d11SharedTextureSource, D3d11SharedTextureSourceError,
+};